+++ asset/source.go
@@ -180,6 +180,7 @@ type ProjectSource struct {
 	Engine            string      `json:"engine"`
 	Entry             string      `json:"entry"`
 	Script            string      `json:"script"`
+	API               int         `json:"api"`
 	Resolution        []int       `json:"resolution"`
 	InspectResolution []int       `json:"inspect_resolution"`
 	TickRate          int         `json:"tick_rate"`